Move CHC trait and remote client to own crate (#4216)
* Fix ignored error

* Move chc_remote to own crate

* Move CHC to own crate

* Clippy

* Module level docs for CHC trait

* Move chc_local to chc crate

* Add ChcHttp docs

* Keep the HTTP impl of chc out if "chc" feature not enabled

* Add test that general CHC failure prevents chain writes

* Re-use reqwest client

* Remove url2 dep
