+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
   "crates/holochain_types",
   "crates/holochain",
   "crates/holochain_cascade",
+  "crates/holochain_chc",
   "crates/holochain_conductor_api",
   "crates/holochain_p2p",
   "crates/holochain_keystore",
