+++ crates/holochain/src/core/sys_validate.rs
@@ -537,71 +537,6 @@ pub fn check_update_reference(
     Ok(())
 }
 
-/// Validate a chain of actions with an optional starting point.
-pub fn validate_chain<'iter, A: 'iter + ChainItem>(
-    mut actions: impl Iterator<Item = &'iter A>,
-    persisted_chain_head: &Option<(A::Hash, u32)>,
-) -> SysValidationResult<()> {
-    // Check the chain starts in a valid way.
-    let mut last_item = match actions.next() {
-        Some(item) => {
-            match persisted_chain_head {
-                Some((prev_hash, prev_seq)) => {
-                    check_prev_action_chain(prev_hash, *prev_seq, item)
-                        .map_err(ValidationOutcome::from)?;
-                }
-                None => {
-                    // If there's no persisted chain head, then the first action
-                    // must have no parent.
-                    if item.prev_hash().is_some() {
-                        return Err(ValidationOutcome::PrevActionError(
-                            (PrevActionErrorKind::InvalidRoot, item).into(),
-                        )
-                        .into());
-                    }
-                }
-            }
-            (item.get_hash(), item.seq())
-        }
-        None => return Ok(()),
-    };
-
-    for item in actions {
-        // Check each item of the chain is valid.
-        check_prev_action_chain(last_item.0, last_item.1, item).map_err(ValidationOutcome::from)?;
-        last_item = (item.get_hash(), item.seq());
-    }
-    Ok(())
-}
-
-// Check the action is valid for the previous action.
-fn check_prev_action_chain<A: ChainItem>(
-    prev_action_hash: &A::Hash,
-    prev_action_seq: u32,
-    action: &A,
-) -> Result<(), PrevActionError> {
-    // The root cannot appear later in the chain
-    if action.prev_hash().is_none() {
-        Err((PrevActionErrorKind::MissingPrev, action).into())
-    } else if action.prev_hash().map_or(true, |p| p != prev_action_hash) {
-        // Check the prev hash matches.
-        Err((PrevActionErrorKind::HashMismatch(action.seq()), action).into())
-    } else if action
-        .seq()
-        .checked_sub(1)
-        .map_or(true, |s| prev_action_seq != s)
-    {
-        // Check the prev seq is one less.
-        Err((
-            PrevActionErrorKind::InvalidSeq(action.seq(), prev_action_seq),
-            action,
-        )
-            .into())
-    } else {
-        Ok(())
-    }
-}
-
 /// Allows DhtOps to be sent to some receiver
 #[async_trait::async_trait]
 #[cfg_attr(test, mockall::automock)]
