+++ crates/holochain/src/core/workflow/sys_validation_workflow/validate_op_tests.rs
@@ -6,7 +6,6 @@ use super::validation_deps::SysValDeps;
 use crate::core::workflow::sys_validation_workflow::types::Outcome;
 use crate::core::workflow::sys_validation_workflow::validate_op;
 use crate::core::workflow::WorkflowResult;
-use crate::core::PrevActionErrorKind;
 use crate::core::ValidationOutcome;
 use crate::prelude::*;
 use ::fixt::prelude::*;
