+++ crates/holochain/src/test_utils/conductor_setup.rs
@@ -12,9 +12,9 @@ use crate::core::ribosome::real_ribosome::RealRibosome;
 use crate::core::ribosome::RibosomeT;
 use holo_hash::AgentPubKey;
 use holo_hash::DnaHash;
+use holochain_chc::ChcImpl;
 use holochain_keystore::MetaLairClient;
 use holochain_p2p::actor::HolochainP2pRefToDna;
-use holochain_p2p::ChcImpl;
 use holochain_p2p::HolochainP2pDna;
 use holochain_serialized_bytes::SerializedBytes;
 use holochain_state::prelude::test_db_dir;
