+++ crates/holochain_cascade/src/test_utils.rs
@@ -8,10 +8,10 @@ use holo_hash::AgentPubKey;
 use holo_hash::AnyDhtHash;
 use holo_hash::AnyDhtHashPrimitive;
 use holo_hash::EntryHash;
+use holochain_chc::ChcImpl;
 use holochain_nonce::Nonce256Bits;
 use holochain_p2p::actor;
 use holochain_p2p::event::CountersigningSessionNegotiationMessage;
-use holochain_p2p::ChcImpl;
 use holochain_p2p::HolochainP2pDnaT;
 use holochain_p2p::HolochainP2pError;
 use holochain_sqlite::rusqlite::Transaction;
