+++ crates/holochain_chc/CHANGELOG.md
@@ -1,0 +1,11 @@
+---
+default_semver_increment_mode: !pre_minor dev
+---
+# Changelog
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## \[Unreleased\]
+
+## 0.1.0-dev.1
+
