+++ crates/holochain_chc/README.md
@@ -1,0 +1,3 @@
+# holochain_chc
+
+License: Apache-2.0
