+++ crates/holochain_p2p/Cargo.toml
@@ -20,6 +20,7 @@ futures = "0.3"
 ghost_actor = "0.3.0-alpha.6"
 hc_sleuth = { version = "^0.4.0-dev.21", path = "../hc_sleuth" }
 holo_hash = { version = "^0.4.0-dev.11", path = "../holo_hash" }
+holochain_chc = { version = "^0.1.0-dev.1", path = "../holochain_chc" }
 holochain_keystore = { version = "^0.4.0-dev.19", path = "../holochain_keystore" }
 holochain_serialized_bytes = "=0.0.55"
 holochain_types = { version = "^0.4.0-dev.21", path = "../holochain_types" }
