+++ crates/holochain_types/Cargo.toml
@@ -132,5 +132,3 @@ sqlite = [
   "holochain_zome_types/sqlite",
   "kitsune_p2p_dht/sqlite",
 ]
-
-chc = []
