+++ crates/holochain_types/src/chain.rs
@@ -14,6 +14,9 @@ use holochain_zome_types::prelude::*;
 #[cfg(all(test, feature = "test_utils"))]
 pub mod test;
 
+mod chain_item;
+pub use chain_item::*;
+
 /// Helpers for constructing AgentActivity
 pub trait AgentActivityExt {
     /// Create an empty chain status
@@ -31,77 +34,6 @@ pub trait AgentActivityExt {
 }
 
 impl AgentActivityExt for AgentActivityResponse {}
-
-/// Abstraction over an item in a chain.
-// Alternate implementations are only used for testing, so this should not
-// add a large monomorphization overhead
-pub trait ChainItem: Clone + PartialEq + Eq + std::fmt::Debug + Send + Sync {
-    /// The type used to represent a hash of this item
-    type Hash: Into<ActionHash>
-        + Clone
-        + PartialEq
-        + Eq
-        + Ord
-        + std::hash::Hash
-        + std::fmt::Debug
-        + Send
-        + Sync;
-
-    /// The sequence in the chain
-    fn seq(&self) -> u32;
-
-    /// The hash of this item
-    fn get_hash(&self) -> &Self::Hash;
-
-    /// The hash of the previous item
-    fn prev_hash(&self) -> Option<&Self::Hash>;
-
-    /// A display representation of the item
-    fn to_display(&self) -> String;
-}
-
-/// Alias for getting the associated hash type of a ChainItem
-pub type ChainItemHash<I> = <I as ChainItem>::Hash;
-
-impl ChainItem for ActionHashed {
-    type Hash = ActionHash;
-
-    fn seq(&self) -> u32 {
-        self.action_seq()
-    }
-
-    fn get_hash(&self) -> &Self::Hash {
-        self.as_hash()
-    }
-
-    fn prev_hash(&self) -> Option<&Self::Hash> {
-        self.prev_action()
-    }
-
-    fn to_display(&self) -> String {
-        format!("{}", self.content)
-    }
-}
-
-impl ChainItem for SignedActionHashed {
-    type Hash = ActionHash;
-
-    fn seq(&self) -> u32 {
-        self.hashed.seq()
-    }
-
-    fn get_hash(&self) -> &Self::Hash {
-        self.hashed.get_hash()
-    }
-
-    fn prev_hash(&self) -> Option<&Self::Hash> {
-        self.hashed.prev_hash()
-    }
-
-    fn to_display(&self) -> String {
-        format!("{}", self.hashed.content)
-    }
-}
 
 #[must_use = "Iterator doesn't do anything unless consumed."]
 #[derive(Debug)]
